Push code complete v1.0

+++ src/main/java/com/laptrinhjavaweb/news/dto/data/BussinessDetails.java
@@ -13,5 +13,5 @@ public class BussinessDetails {
     String accountCode;
     String bussinessRegNo;
     String companyRegNo;
-    Float taxRate;
+    Double taxRate;
 }
