+++ src/main/java/com/laptrinhjavaweb/news/dto/request/mongo/CouponInput.java
@@ -1,5 +1,7 @@
 package com.laptrinhjavaweb.news.dto.request.mongo;
 
+import java.math.BigDecimal;
+
 import lombok.Data;
 
 @Data
@@ -7,9 +9,10 @@ public class CouponInput {
     private String _id;
     private String id;
     private String title;
-    private double discount;
+    private BigDecimal discount;
     private boolean enabled;
-    public String getId(){
+
+    public String getId() {
         return _id;
     }
 }
