+++ src/main/java/com/laptrinhjavaweb/news/dto/request/mongo/VariationInput.java
@@ -1,14 +1,15 @@
 package com.laptrinhjavaweb.news.dto.request.mongo;
 
-import lombok.Data;
-
+import java.math.BigDecimal;
 import java.util.List;
+
+import lombok.Data;
 
 @Data
 public class VariationInput {
     private String title;
     private List<String> addons;
     private boolean isOutOfStock;
-    private Float price;
-    private Float discounted;
+    private BigDecimal price;
+    private BigDecimal discounted;
 }
