+++ src/main/java/com/laptrinhjavaweb/news/mongo/CouponDocument.java
@@ -1,11 +1,13 @@
 package com.laptrinhjavaweb.news.mongo;
 
+import org.bson.types.Decimal128;
+import org.springframework.data.annotation.Id;
+import org.springframework.data.mongodb.core.mapping.Document;
+
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.springframework.data.annotation.Id;
-import org.springframework.data.mongodb.core.mapping.Document;
 
 @Document("coupon")
 @AllArgsConstructor
@@ -15,10 +17,12 @@ import org.springframework.data.mongodb.core.mapping.Document;
 public class CouponDocument {
     @Id
     private String id;
+
     private String title;
-    private double discount;
+    private Decimal128 discount;
     private boolean enabled;
-    private String get_id(){
+
+    private String get_id() {
         return id;
     }
 }
